snyk fixes. Refactor and cleanup a little
