+++ view-creator-framework/src/test/resources/sample-view-generation-spec.conf
@@ -1,16 +1,10 @@
 tool.class = org.hypertrace.core.viewcreator.pinot.PinotTableCreationTool
 service.name = "sample-view"
+
 view = {
   name = myView
   output.schema.class = org.hypertrace.core.viewcreator.test.api.TestView
   output.schema.url = "host:port/myView"
-}
-
-kafka = {
-  brokerAddress = "localhost:9092"
-  topicName = test-view-events
-  partitions = 8
-  replicationFactor = 3
 }
 
 pinot = {
@@ -48,5 +42,11 @@ pinot = {
     realtime.segment.flush.threshold.time: 3600000,
     realtime.segment.flush.threshold.size: 500000,
     stream.kafka.consumer.prop.auto.offset.reset: largest
+  }
+
+  testProps = {
+    key1: value1
+    key2.1: values
+    key3.1.1: values
   }
 }
